+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 - Mermaid 8.11.0 support
 - D3 7.0.0 support
 - Remove issue macro, which is now supported by Redmine core
+- new ticket message can be overwritten for projects
 
 3.0.2
 +++++